rename ethidx to ethindex
as requested by bernd.
