+++ print-logs
@@ -2,7 +2,7 @@
 
 import json
 from web3 import Web3
-from ethidx import logdecode
+from ethindex import logdecode
 
 
 def main():
